+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [0.8.3]
+
+### _Feature_
+
+- `Persistent.equals`, `Persistent.notEquals`, `Persistent.assign` now support function evaluator as argument
+- Added `hidden` and `disabled` config to `Menu` choice
+- Added `Menu.hideIf` and `Menu.disableIf` magic methods
+
 ## [0.8.2]
 
 ### Fixed